requirements.txt
fastapi
uvicorn
redis
python-dotenv
psycopg2-binary
SQLAlchemy
